+++ app/Http/Requests/DepartureRelease/AcknowledgeDepartureRelease.php
@@ -5,7 +5,7 @@ namespace App\Http\Requests\DepartureRelease;
 use App\Rules\Controller\ControllerPositionValid;
 use Illuminate\Foundation\Http\FormRequest;
 
-class AcknowledgeDepartureReleaseRequest extends FormRequest
+class AcknowledgeDepartureRelease extends FormRequest
 {
     public function rules()
     {
