+++ app/Http/Requests/DepartureRelease/ApproveDepartureRelease.php
@@ -5,7 +5,7 @@ namespace App\Http\Requests\DepartureRelease;
 use App\Rules\Controller\ControllerPositionValid;
 use Illuminate\Foundation\Http\FormRequest;
 
-class ApproveDepartureReleaseRequest extends FormRequest
+class ApproveDepartureRelease extends FormRequest
 {
     public function rules()
     {
